Move backend code and add Express
All backend code is moved to a lower directory in order to clean up the
file structure. There was too much junk at the root directory.

Express is added to the project. The node server must run on a port
different than usual to prevent port conflicts between React and Node
running on the same local machine.

An example GET endpoint is supplied for the client to fetch.